+++ cloud/auth.go
@@ -142,6 +142,59 @@ func (c *Client) DeleteAuthCache(ctx context.Context) error {
 	return nil
 }
 
+func (c *Client) DisableAutoLogin(ctx context.Context) error {
+	path, err := c.getDisableAutoLoginPath(true)
+	if err != nil {
+		return errors.Wrapf(err, "failed to get disable auto login path")
+	}
+	f, err := os.Create(path)
+	if err != nil {
+		return errors.Wrapf(err, "failed to create %s", path)
+	}
+	err = f.Close()
+	if err != nil {
+		return errors.Wrapf(err, "failed to close %s", path)
+	}
+	return nil
+}
+
+func (c *Client) IsAutoLoginPermitted(ctx context.Context) (bool, error) {
+	path, err := c.getDisableAutoLoginPath(true)
+	if err != nil {
+		return false, errors.Wrapf(err, "failed to get disable auto login path")
+	}
+	info, err := os.Stat(path)
+	if err != nil {
+		if os.IsNotExist(err) {
+			return true, nil // auto login is allowed when the file does not exist
+		}
+		return false, err
+	}
+	if info.IsDir() {
+		return false, fmt.Errorf("expected %s to be a file (not a directory)", path)
+	}
+	return false, nil
+}
+
+func (c *Client) EnableAutoLogin(ctx context.Context) error {
+	canAutoLogin, err := c.IsAutoLoginPermitted(ctx)
+	if err != nil {
+		return errors.Wrapf(err, "failed to determine if auto login is permitted")
+	}
+	if canAutoLogin {
+		return nil // already allowed
+	}
+	path, err := c.getDisableAutoLoginPath(true)
+	if err != nil {
+		return errors.Wrapf(err, "failed to get disable auto login path")
+	}
+	err = os.Remove(path)
+	if err != nil {
+		return errors.Wrapf(err, "failed to remove %s", path)
+	}
+	return nil
+}
+
 func (c *Client) SetTokenCredentials(ctx context.Context, token string) (string, error) {
 	c.email = ""
 	c.password = ""
@@ -252,23 +305,6 @@ func (c *Client) loadToken() error {
 	return nil
 }
 
-func (c *Client) getCredentialsPath(create bool) (string, error) {
-	confDirPath := c.authDir
-	if confDirPath == "" {
-		if create {
-			var err error
-			confDirPath, err = cliutil.GetOrCreateEarthlyDir(c.installationName)
-			if err != nil {
-				return "", errors.Wrap(err, "cannot get .earthly dir")
-			}
-		} else {
-			confDirPath = cliutil.GetEarthlyDir(c.installationName)
-		}
-	}
-	credPath := filepath.Join(confDirPath, "auth.credentials")
-	return credPath, nil
-}
-
 func (c *Client) migrateOldToken() error {
 	confDirPath := c.authDir
 	if confDirPath == "" {
@@ -284,10 +320,22 @@ func (c *Client) migrateOldToken() error {
 	return nil
 }
 
+func (c *Client) getCredentialsPath(create bool) (string, error) {
+	return c.getAuthPath("auth.credentials", create)
+}
+
 func (c *Client) getTokenPath(create bool) (string, error) {
+	return c.getAuthPath("auth.jwt", create)
+}
+
+func (c *Client) getDisableAutoLoginPath(create bool) (string, error) {
+	return c.getAuthPath("do-not-login-automatically", create)
+}
+
+func (c *Client) getAuthPath(filename string, createEarthlyDir bool) (string, error) {
 	confDirPath := c.authDir
 	if confDirPath == "" {
-		if create {
+		if createEarthlyDir {
 			var err error
 			confDirPath, err = cliutil.GetOrCreateEarthlyDir(c.installationName)
 			if err != nil {
@@ -297,7 +345,7 @@ func (c *Client) getTokenPath(create bool) (string, error) {
 			confDirPath = cliutil.GetEarthlyDir(c.installationName)
 		}
 	}
-	tokenPath := filepath.Join(confDirPath, "auth.jwt")
+	tokenPath := filepath.Join(confDirPath, filename)
 	return tokenPath, nil
 }
 
@@ -402,7 +450,11 @@ func (c *Client) loginWithToken(ctx context.Context) error {
 }
 
 func (c *Client) loginWithSSH(ctx context.Context) error {
-	if c.disableSSHKeyGuessing {
+	allowAutoLogin, err := c.IsAutoLoginPermitted(ctx)
+	if err != nil {
+		return err
+	}
+	if c.disableSSHKeyGuessing || !allowAutoLogin {
 		return ErrNoAuthorizedPublicKeys
 	}
 	challenge, err := c.getChallenge(ctx)