+++ cmd/earthly/account_cmds.go
@@ -161,7 +161,7 @@ func (app *earthlyApp) accountCmds() []*cli.Command {
 		},
 		{
 			Name:  "reset",
-			Usage: "Reset Earthly account password *beta*",
+			Usage: "Reset Earthly account password",
 			UsageText: "earthly [options] account reset --email <email>\n" +
 				"   earthly [options] account reset --email <email> --token <token>\n",
 			Action: app.actionAccountReset,
@@ -523,6 +523,18 @@ func (app *earthlyApp) actionAccountLogin(cliCtx *cli.Context) error {
 	token := app.token
 	pass := app.password
 
+	cloudClient, err := app.newCloudClient()
+	if err != nil {
+		return err
+	}
+
+	// if a user explicitly calls Login, we will allow future auto-logins
+	// which is required for refreshing tokens
+	err = cloudClient.EnableAutoLogin(cliCtx.Context)
+	if err != nil {
+		return errors.Wrap(err, "failed to enable auto login")
+	}
+
 	if cliCtx.NArg() == 1 {
 		emailOrToken := cliCtx.Args().First()
 		if token == "" && email == "" {
@@ -541,10 +553,6 @@ func (app *earthlyApp) actionAccountLogin(cliCtx *cli.Context) error {
 
 	if token != "" && (email != "" || pass != "") {
 		return errors.New("--token cannot be used in conjuction with --email or --password")
-	}
-	cloudClient, err := app.newCloudClient()
-	if err != nil {
-		return err
 	}
 
 	// special case where global auth token overrides login logic
@@ -690,6 +698,10 @@ func (app *earthlyApp) actionAccountLogout(cliCtx *cli.Context) error {
 	err = cloudClient.DeleteAuthCache(cliCtx.Context)
 	if err != nil {
 		return errors.Wrap(err, "failed to logout")
+	}
+	err = cloudClient.DisableAutoLogin(cliCtx.Context)
+	if err != nil {
+		return errors.Wrap(err, "failed to disable auto login")
 	}
 	return nil
 }