+++ cmd/earthly/buildkit.go
@@ -113,11 +113,11 @@ func (app *earthlyApp) configureSatellite(cliCtx *cli.Context, cloudClient *clou
 		app.buildkitdSettings.ServerTLSKey = ""
 	}
 
-	orgID, err := app.getSatelliteOrgID(cliCtx.Context, cloudClient)
+	orgName, orgID, err := app.getSatelliteOrg(cliCtx.Context, cloudClient)
 	if err != nil {
 		return errors.Wrap(err, "failed getting org")
 	}
-	satelliteName, err := app.getSatelliteName(cliCtx.Context, orgID, app.satelliteName, cloudClient)
+	satelliteName, err := app.getSatelliteName(cliCtx.Context, orgName, app.satelliteName, cloudClient)
 	if err != nil {
 		return errors.Wrap(err, "failed getting satellite name")
 	}
@@ -145,7 +145,7 @@ func (app *earthlyApp) configureSatellite(cliCtx *cli.Context, cloudClient *clou
 
 	// Reserve the satellite for the upcoming build.
 	// This operation can take a moment if the satellite is asleep.
-	err = app.reserveSatellite(cliCtx.Context, cloudClient, satelliteName, app.satelliteName, orgID, gitAuthor, gitConfigEmail)
+	err = app.reserveSatellite(cliCtx.Context, cloudClient, satelliteName, app.satelliteName, orgName, gitAuthor, gitConfigEmail)
 	if err != nil {
 		return err
 	}
@@ -165,10 +165,10 @@ func (app *earthlyApp) isUsingSatellite(cliCtx *cli.Context) bool {
 	return app.cfg.Satellite.Name != "" || app.satelliteName != ""
 }
 
-func (app *earthlyApp) reserveSatellite(ctx context.Context, cloudClient *cloud.Client, name, displayName, orgID, gitAuthor, gitConfigEmail string) error {
+func (app *earthlyApp) reserveSatellite(ctx context.Context, cloudClient *cloud.Client, name, displayName, orgName, gitAuthor, gitConfigEmail string) error {
 	console := app.console.WithPrefix("satellite")
 	_, isCI := analytics.DetectCI(app.earthlyCIRunner)
-	out := cloudClient.ReserveSatellite(ctx, name, orgID, gitAuthor, gitConfigEmail, isCI)
+	out := cloudClient.ReserveSatellite(ctx, name, orgName, gitAuthor, gitConfigEmail, isCI)
 	err := showSatelliteLoading(console, displayName, out)
 	if err != nil {
 		return errors.Wrap(err, "failed reserving satellite for build")