+++ tests/bootstrap/test-bootstrap.sh
@@ -12,12 +12,15 @@ earthly=$(realpath "$earthly")
 frontend="${frontend:-$(which docker || which podman)}"
 test -n "$frontend" || (>&2 echo "Error: frontend is empty" && exit 1)
 
+# so that we can test production/staging binaries
+default_install_name="${default_install_name:-"earthly-dev"}"
+
 echo "=== Test 1: Hand Bootstrapped ==="
 
 "$earthly" bootstrap
 
-if [[ ! -d "$HOME/.earthly-dev" ]]; then
-  echo ".earthly-dev directory was missing after bootstrap"
+if [[ ! -d "$HOME/.${default_install_name}" ]]; then
+  echo ".${default_install_name} directory was missing after bootstrap"
   exit 1
 fi
 
@@ -36,14 +39,14 @@ if  cat hand_boot_output | grep -q "bootstrap |"; then
     exit 1
 fi
 
-rm -rf "$HOME/.earthly-dev" "$HOME/.earthly-test"
+rm -rf "$HOME/.${default_install_name}" "$HOME/.earthly-test"
 
 echo "=== Test 2: Implied Bootstrap ==="
 
 "$earthly" +test
 
-if [[ ! -d "$HOME/.earthly-dev" ]]; then
-  echo ".earthly-dev directory was missing after bootstrap"
+if [[ ! -d "$HOME/.${default_install_name}" ]]; then
+  echo ".${default_install_name} directory was missing after bootstrap"
   exit 1
 fi
 
@@ -62,21 +65,21 @@ if  cat imp_boot_output | grep -q "bootstrap |"; then
     exit 1
 fi
 
-rm -rf "$HOME/.earthly-dev" "$HOME/.earthly-test"
+rm -rf "$HOME/.${default_install_name}" "$HOME/.earthly-test"
 
 echo "=== Test 3: CI ==="
 
 "$earthly" --ci +test
 
-if [[ ! -d "$HOME/.earthly-dev" ]]; then
-  echo ".earthly-dev directory was missing after bootstrap"
+if [[ ! -d "$HOME/.${default_install_name}" ]]; then
+  echo ".${default_install_name} directory was missing after bootstrap"
   exit 1
 fi
 
 EARTHLY_INSTALLATION_NAME=earthly-test "$earthly" --ci +test
 
 if [[ ! -d "$HOME/.earthly-test" ]]; then
- echo ".earthly-dev directory was missing after bootstrap"
+ echo ".earthly-test directory was missing after bootstrap"
  exit 1
 fi
 
@@ -88,7 +91,7 @@ if  cat ci_boot_output | grep -q "bootstrap |"; then
     exit 1
 fi
 
-rm -rf "$HOME/.earthly-dev" "$HOME/.earthly-test"
+rm -rf "$HOME/.${default_install_name}" "$HOME/.earthly-test"
 
 echo "=== Test 4: With Autocomplete ==="
 
@@ -107,7 +110,7 @@ if [[ ! -f "/usr/share/bash-completion/completions/earthly" ]]; then
   exit 1
 fi
 
-rm -rf "$HOME/.earthly-dev"
+rm -rf "$HOME/.${default_install_name}"
 sudo rm -rf "/usr/share/bash-completion/completions/earthly"
 
 echo "=== Test 5: Permissions ==="
@@ -122,46 +125,46 @@ echo "Group: $GRP"
 
 "$earthly" bootstrap
 
-if [[ $(stat --format '%U' "$HOME/.earthly-dev") != "$USR" ]]; then
+if [[ $(stat --format '%U' "$HOME/.${default_install_name}") != "$USR" ]]; then
   echo "earthly directory is not owned by the user"
-  stat "$HOME/.earthly-dev"
+  stat "$HOME/.${default_install_name}"
   exit 1
 fi
 
-if [[ $(stat --format '%G' "$HOME/.earthly-dev") != "$GRP" ]]; then
+if [[ $(stat --format '%G' "$HOME/.${default_install_name}") != "$GRP" ]]; then
   echo "earthly directory is not owned by the users group"
-  stat "$HOME/.earthly-dev"
+  stat "$HOME/.${default_install_name}"
   exit 1
 fi
 
 echo "----"
 
-touch $HOME/.earthly-dev/config.yml
-sudo chown -R 12345:12345 $HOME/.earthly-dev
+touch $HOME/.${default_install_name}/config.yml
+sudo chown -R 12345:12345 $HOME/.${default_install_name}
 
 sudo "$earthly" bootstrap
 
-if [[ $(stat --format '%U' "$HOME/.earthly-dev") != "$USR" ]]; then
+if [[ $(stat --format '%U' "$HOME/.${default_install_name}") != "$USR" ]]; then
   echo "earthly directory is not owned by the user"
-  stat "$HOME/.earthly-dev"
+  stat "$HOME/.${default_install_name}"
   exit 1
 fi
 
-if [[ $(stat --format '%G' "$HOME/.earthly-dev") != "$GRP" ]]; then
+if [[ $(stat --format '%G' "$HOME/.${default_install_name}") != "$GRP" ]]; then
   echo "earthly directory is not owned by the users group"
-  stat "$HOME/.earthly-dev"
+  stat "$HOME/.${default_install_name}"
   exit 1
 fi
 
-if [[ $(stat --format '%U' "$HOME/.earthly-dev/config.yml") != "$USR" ]]; then
+if [[ $(stat --format '%U' "$HOME/.${default_install_name}/config.yml") != "$USR" ]]; then
   echo "earthly config is not owned by the user"
-  stat "$HOME/.earthly-dev/config.yml"
+  stat "$HOME/.${default_install_name}/config.yml"
   exit 1
 fi
 
-if [[ $(stat --format '%G' "$HOME/.earthly-dev/config.yml") != "$GRP" ]]; then
+if [[ $(stat --format '%G' "$HOME/.${default_install_name}/config.yml") != "$GRP" ]]; then
   echo "earthly config is not owned by the users group"
-  stat "$HOME/.earthly-dev/config.yml"
+  stat "$HOME/.${default_install_name}/config.yml"
   exit 1
 fi
 
@@ -217,7 +220,7 @@ if ! DOCKER_HOST="$frontend is missing" "$earthly" bootstrap --source zsh > /dev
   exit 1
 fi
 
-rm -rf "$HOME/.earthly-dev"
+rm -rf "$HOME/.${default_install_name}"
 
 echo "=== Test 8: No Buildkit ==="
 
@@ -232,4 +235,4 @@ if ! DOCKER_HOST="$frontend is missing" "$earthly" bootstrap --no-buildkit; then
   exit 1
 fi
 
-rm -rf "$HOME/.earthly-dev"
+rm -rf "$HOME/.${default_install_name}"