+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to [Earthly](https://github.com/earthly/earthly) will be doc
 
 ## Unreleased
 
+## v0.7.9 - 2023-06-22
+
+### Changed
+- The command `docker-build` now also supports passing multiple platforms using a comma (e.g `--platform linux/amd64,linux/arm64`)
+- Increased temporary lease duration of buildkit's history queue to prevent unknown history in blob errors under high cpu load. [#3000](https://github.com/earthly/earthly/issues/3000)
+- Performing an `earthly account logout` will keep you logged out -- earthly will no longer attempt an auto-login via ssh-agent (use `earthly account login` to log back in).
+
+### Fixed
+- Fixed a bug in satellite update command which was incorrectly changing satellites to medium size.
+- Fixed support for being authenticated with multiple registries when using the cloud-based `earthly registry` feature. [#3010](https://github.com/earthly/earthly/issues/3010)
+- Fixed `WITH DOCKER` auto install script when using latest (bookworm) version.
+
+### Added
+- Buildkit logs now include version and revision.
+- Satellite name autocompletion
+
 ## v0.7.8 - 2023-06-07
 
 ### Added